+++ frontend/src/routes/PrivateRoute.tsx
@@ -2,7 +2,7 @@ import { Routes, Route, useLocation, useNavigate } from 'react-router-dom'
 import Profile from '../pages/Profile'
 import ScanTask from '../pages/TransportWorker/ScanQRCode'
 import Cart from '../pages/TransportWorker/Cart'
-import UnloadSuccess from '../pages/TransportWorker/Success'
+import UnloadSuccess from '../pages/Success'
 import { useContext, useEffect } from 'react'
 import { AuthContext } from '../contexts/auth'
 import Dashboard from '../pages/Dashboard'
